Possible implementation, however...:
The RFCs new definitions are poorly done and redundant.
  address = mailbox / group
So:
  (mailbox / address) = (mailbox / (mailbox / group))   <--- redundant!
Yet they define
  sender = "Sender:" (mailbox / address) CRLF
And
  from = "From:" (mailbox-list / address-list) CRLF

They should have written:
  sender = "Sender:" address CRLF
  from = "From:" address-list CRLF

The code in this commit is always using the Address form for Sender and From,
never the Mailbox form (it has degenerate enum variants that never get used) because
I coded this literally to what the RFC said.